+++ lib/ftp/fnFTP.h
@@ -133,7 +133,7 @@ public:
      * @param path path to file
      * @return size of file in bytes, or -1 on error.
      */
-    size_t get_file_size(string path);
+    int32_t get_file_size(string path);
 
 protected:
 private:
